chore: api 함수 수정
