+++ src/shared/lib/api/index.ts
@@ -1,3 +1,10 @@
-export { clientApi } from './api'
+export { clientApi } from './clientApi'
 export { serverApi } from './serverApi'
-export type { ApiError, NetworkError, ApiResult, ApiResponse, ApiFieldError, RequestConfig } from './types'
+export type {
+  ApiError,
+  NetworkError,
+  ApiResult,
+  ApiResponse,
+  ApiFieldError,
+  RequestConfig,
+} from './types'
